requirements.txt
flet>=0.80
pyserial>=3.5
